Fix AXChildren and AXContents of tab bar (fixes New Tab button)
AXChildren should return all (non-ignored) subviews, and now that OakTabBarView
and all elements inside it (tabs and buttons) are view-based, we can leave
AppKit to do the work (no need to override the default implementation of the
attribute).  This also fixes the New Tab button which is now available for
VoiceOver (though VoiceOver users will much more probably use Option-Cmd-N to
be more effective).

AXContents should return in our case the same as AXChildren does.

AXTabs are left as they are to return only the tabs (e.g. not to return the New
Tab button).

This was also discussed in textmate/bugs#7.